requirements.txt
certifi==2025.8.3
elastic-transport==9.1.0
elasticsearch==9.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
six==1.17.0
typing_extensions==4.15.0
urllib3==2.5.0
